requirements.txt
google-generativeai
pyyaml
tqdm
python-dotenv
